fix: room export by email
