+++ packages/rest-typings/src/v1/rooms/index.ts
@@ -1,0 +1,2 @@
+export * from './rooms';
+export * from './RoomsExportProps';
